Whitelist strprintf::fmt()'s input types
Up until now, strprintf::fmt() gladly accepted anything that has been
thrown at it: ints, strings, void pointers, std::vector iterators. It's
the latter that cause problems: sprintf (used by fmt() internally)
doesn't know how to format these complicated objects, and prints some
garbage instead. C++ didn't notice, but as I work on the new Rust
logger, I found that some log messages aren't valid UTF-8 because of
this.

The only solution I could come up with is to whitelist the input types.
strprintf::fmt() will fail the compilation with a long scary template
error if you feed it something it doesn't expect. I don't know how to
make the error any better, but hopefully it won't occur that often, and
is simple enough to understand if you read it from the beginning.

+++ src/pbcontroller.cpp
@@ -231,7 +231,7 @@ int PbController::run(int argc, char* argv[])
 				std::cerr << strprintf::fmt(_("%s: %d: invalid "
 							      "loglevel value"),
 						     argv[0],
-						     l)
+						     static_cast<int>(l))
 					  << std::endl;
 				return EXIT_FAILURE;
 			}
